refactor: Repository를 각각의 뷰모델에 맞춰 분리
